requirements.txt
django
python-decouple
pyrebase4
django-cors-headers